Update for formal-background.tex
The system does not let me commit the current version of formal-background.tex, so I'm trying this workaround by providing a new file - please take it and rename into the original name. I'd really like my current draft to be available to all authors for use and feedback.
There might be issues with my use of localbibliography.bib on my local copy relative to the master branch. How can they be resolved? Should I propose creating a "richter.bib"?
Is there a more direct way to do all this?