Fix externalization.
The externalization works from both "main.tex", and "chapters/<chapter>.tex".
Even more, they share the externalized files.

This feature, accessed by Forest keys "external/master" and "external/master
dir", requires a so-far unreleased version of Forest; the upcoming "forest.sty"
is therefore included in the commit (in the main folder, and linked into the
"chapters" folder).  The externalized files produced by compilation of either
the master file or the chapter files are saved to folder
"hpsg-handbook.for.dir"; the control file is "hpsg-handbook.for(.tmp)".
".gitignore" and Makefile rule cleaning rules have been updated accordingly.

"\tikzexternalize" is called from compilable ".tex" files themselves, so that
the externalization can be enabled or disabled on per-file basis.

The following issues were fixed to make externalization work:
* The "\renewindex{default}..." problem, discussed in
https://tex.stackexchange.com/questions/352157/tikz-externalization-and-renewindex/352160,
but with a new solution.
* Externalization is disabled during the calls of "\todo" (todonotes), as
recommended in the docs of the package.
* There is a problem with some trees containing AVM's "\@" command. It will be
investigated, but for now, the externalization of those trees is simply disabled.